pyproject.toml
#This file is used to package the InvenTreeReportLsp plugin.
#
#- It was generated by the InvenTree Plugin Creator tool - version 1.13.0
#- Ref: https://github.com/inventree/plugin_creator

[build-system]
requires = ["setuptools", "twine", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "inventree-report-lsp-plugin"
description = "A report editor plugin that is powered by an LSP"
dynamic = ["version"]
authors = [
    { name = "wolflu05",email = "76838159+wolflu05@users.noreply.github.com" },
]
readme = "README.md"
license = "GPL-3.0+"
keywords = ["inventree", "plugin"]
requires-python = ">=3.9"
dependencies = [
    # Enter your plugin library dependencies here
    "websockets==15.0.1",
    "pyjwt==2.10.1",
    "django-template-lsp==1.2.0",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Framework :: InvenTree",
]
[project.urls]
Homepage = "https://github.com/wolflu05/inventree-report-lsp-plugin"

[project.entry-points."inventree_plugins"]
reporteditor = "report_lsp.core:ReportEditor"

[project.entry-points."console_scripts"]
inventree-report-lsp = "report_lsp.lsp.server:cli"

[tool.setuptools.packages.find]
where = [""]

[tool.setuptools.dynamic]
version = {attr = "report_lsp.PLUGIN_VERSION"} 

[tool.ruff]
exclude = [
    ".git",
    "__pycache__",
    "test.py",
    "tests",
    "venv",
    "env",
    ".venv",
    ".env",
]

src=["report_lsp"]
